PP-6831 Rename email model to email.service
This class is a service rather than a model, so rename and refactor to
reflect that.
