+++ app/controllers/email-notifications/email-notifications.controller.js
@@ -6,7 +6,7 @@ const _ = require('lodash')
 // Local dependencies
 const logger = require('../../utils/logger')(__filename)
 const response = require('../../utils/response.js').response
-const Email = require('../../models/email.js')
+const emailService = require('../../services/email.service.js')
 const paths = require('../../paths.js')
 const CORRELATION_HEADER = require('../../utils/correlation-header.js').CORRELATION_HEADER
 
@@ -36,8 +36,7 @@ const showRefundEmail = function (req, res, resource, locals) {
 const toggleConfirmationEmail = function (req, res, enabled) {
   const accountID = req.account.gateway_account_id
   const correlationId = _.get(req, 'headers.' + CORRELATION_HEADER, '')
-  const emailModel = Email(correlationId)
-  emailModel.setConfirmationEnabled(accountID, enabled)
+  emailService.setConfirmationEnabled(accountID, enabled, correlationId)
     .then(() => {
       logger.info(`[${correlationId}] - Updated confirmation email enabled(${enabled}). user=${req.session.passport.user}, gateway_account=${accountID}`)
       res.redirect(303, indexPath)
@@ -59,8 +58,7 @@ module.exports.collectionEmailUpdate = (req, res) => {
   const emailCollectionMode = req.body['email-collection-mode']
   const accountID = req.account.gateway_account_id
   const correlationId = _.get(req, 'headers.' + CORRELATION_HEADER, '')
-  const emailModel = Email(correlationId)
-  emailModel.setEmailCollectionMode(accountID, emailCollectionMode)
+  emailService.setEmailCollectionMode(accountID, emailCollectionMode, correlationId)
     .then(() => {
       logger.info(`[${correlationId}] - Updated email collection mode (${emailCollectionMode}). user=${req.session.passport.user}, gateway_account=${accountID}`)
       res.redirect(303, indexPath)
@@ -102,8 +100,7 @@ module.exports.refundEmailUpdate = (req, res) => {
   const emailRefundEnabled = req.body['email-refund-enabled'] === 'true'
   const accountID = req.account.gateway_account_id
   const correlationId = _.get(req, 'headers.' + CORRELATION_HEADER, '')
-  const emailModel = Email(correlationId)
-  emailModel.setRefundEmailEnabled(accountID, emailRefundEnabled)
+  emailService.setRefundEmailEnabled(accountID, emailRefundEnabled, correlationId)
     .then(() => {
       logger.info(`[${correlationId}] - Updated refund email enabled(${emailRefundEnabled}). user=${req.session.passport.user}, gateway_account=${accountID}`)
       res.redirect(303, indexPath)
@@ -150,8 +147,7 @@ module.exports.update = (req, res) => {
   const newEmailText = req.body['custom-email-text']
   const accountID = req.account.gateway_account_id
   const correlationId = _.get(req, 'headers.' + CORRELATION_HEADER, '')
-  const emailModel = Email(correlationId)
-  emailModel.updateConfirmationTemplate(accountID, newEmailText)
+  emailService.updateConfirmationTemplate(accountID, newEmailText, correlationId)
     .then(() => {
       logger.info(`[${correlationId}] - Updated email notifications custom paragraph. user=${req.session.passport.user}, gateway_account=${accountID}`)
       res.redirect(303, indexPath)
