+++ app/middleware/get-email-notification.js
@@ -1,11 +1,12 @@
 'use strict'
 const { renderErrorView } = require('../utils/response.js')
-const Email = require('../models/email.js')
+const { getEmailSettings } = require('../services/email.service.js')
 const _ = require('lodash')
 const CORRELATION_HEADER = require('../utils/correlation-header.js').CORRELATION_HEADER
 
 module.exports = function (req, res, next) {
-  return Email(req.headers[CORRELATION_HEADER]).get(req.account.gateway_account_id)
+  const correlationId = req.headers[CORRELATION_HEADER]
+  return getEmailSettings(req.account.gateway_account_id, correlationId)
     .then(emailData => {
       req.account = _.merge(req.account, emailData)
       next()
