+++ test/unit/services/email.service.test.js
@@ -7,27 +7,27 @@ const _ = require('lodash')
 const EventEmitter = require('events').EventEmitter
 const proxyquire = require('proxyquire')
 
+function getEmailService (connectorClientStub) {
+  return proxyquire(path.join(__dirname, '/../../../app/services/email.service.js'), {
+    '../services/clients/connector.client.js': connectorClientStub
+  })
+}
+
 describe('email notification', function () {
   describe('getting the template body', function () {
     describe('when connector returns an error', function () {
       it('should return client unavailable', function () {
-        // Create a class that inherits from EventEmitter and emit a 'connectorError' event which is handled by the service
-        class StubConnectorEmailFunctions {
-          getAccount () {
-            return Promise.reject(new Error('connection error'))
-          }
-        }
-        let sCEFinst = new StubConnectorEmailFunctions()
-        let connectorClientStub = {
+        const connectorClientStub = {
           ConnectorClient: function () {
-            return sCEFinst
+            return {
+              getAccount: function () {
+                return Promise.reject(new Error('connection error'))
+              }
+            }
           }
         }
-        let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-          '../services/clients/connector.client.js': connectorClientStub
-        })
-        const emailModel = Email('some-unique-id')
-        return expect(emailModel.get(123))
+        const emailService = getEmailService(connectorClientStub)
+        return expect(emailService.getEmailSettings(123, 'some-unique-id'))
           .to.be.rejectedWith('CONNECTOR_FAILED')
       }
       )
@@ -35,42 +35,37 @@ describe('email notification', function () {
 
     describe('when connector returns correctly', function () {
       it('should return the correct promise', function () {
-        class StubConnectorEmailFunctions {
-          getAccount (params) {
-            /* eslint-disable */
-            return Promise.resolve(
-              {
-                gateway_account_id: 31,
-                service_name: '8b9370c1a83c4d71a538a1691236acc2',
-                type: 'test',
-                analytics_id: '8b02c7e542e74423aa9e6d0f0628fd58',
-                email_collection_mode: 'MANDATORY',
-                email_notifications: {
-                  PAYMENT_CONFIRMED: {
-                    version: 1,
-                    enabled: true,
-                    template_body: 'template here'
-                  },
-                  REFUND_ISSUED: {
-                    version: 1,
-                    enabled: true
-                  }
-                }
-              })
-            /* eslint-enable */
-          }
-        }
-        let sCEFinst = new StubConnectorEmailFunctions()
-        let connectorClientStub = {
+        const connectorClientStub = {
           ConnectorClient: function () {
-            return sCEFinst
+            return {
+              getAccount: function () {
+                /* eslint-disable */
+                return Promise.resolve(
+                  {
+                    gateway_account_id: 31,
+                    service_name: '8b9370c1a83c4d71a538a1691236acc2',
+                    type: 'test',
+                    analytics_id: '8b02c7e542e74423aa9e6d0f0628fd58',
+                    email_collection_mode: 'MANDATORY',
+                    email_notifications: {
+                      PAYMENT_CONFIRMED: {
+                        version: 1,
+                        enabled: true,
+                        template_body: 'template here'
+                      },
+                      REFUND_ISSUED: {
+                        version: 1,
+                        enabled: true
+                      }
+                    }
+                  })
+                /* eslint-enable */
+              }
+            }
           }
         }
-        let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-          '../services/clients/connector.client.js': connectorClientStub
-        })
-        const emailModel = Email('some-unique-id')
-        return expect(emailModel.get(123))
+        const emailService = getEmailService(connectorClientStub)
+        return expect(emailService.getEmailSettings(123, 'some-unique-id'))
           .to.be.fulfilled.then(function (response) {
             expect(response).to.deep.equal({
               customEmailText: 'template here',
@@ -101,11 +96,8 @@ describe('email notification', function () {
             return sCEFinst
           }
         }
-        let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-          '../services/clients/connector.client.js': connectorClientStub
-        })
-        const emailModel = Email('some-unique-id')
-        return expect(emailModel.updateConfirmationTemplate(123))
+        const emailService = getEmailService(connectorClientStub)
+        return expect(emailService.updateConfirmationTemplate(123, 'some-unique-id'))
           .to.be.rejectedWith('CONNECTOR_FAILED')
       }
       )
@@ -128,11 +120,8 @@ describe('email notification', function () {
             return sCEFinst
           }
         }
-        let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-          '../services/clients/connector.client.js': connectorClientStub
-        })
-        const emailModel = Email('some-unique-id')
-        return expect(emailModel.updateConfirmationTemplate(123))
+        const emailService = getEmailService(connectorClientStub)
+        return expect(emailService.updateConfirmationTemplate(123, 'some-unique-id'))
           .to.be.rejectedWith('POST_FAILED')
       })
     })
@@ -151,11 +140,8 @@ describe('email notification', function () {
             return sCEFinst
           }
         }
-        let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-          '../services/clients/connector.client.js': connectorClientStub
-        })
-        const emailModel = Email('some-unique-id')
-        return expect(emailModel.updateConfirmationTemplate(123)).to.be.fulfilled
+        const emailService = getEmailService(connectorClientStub)
+        return expect(emailService.updateConfirmationTemplate(123, 'some-unique-id')).to.be.fulfilled
       })
     })
   })
@@ -179,11 +165,8 @@ describe('email notification', function () {
               return sCEFinst
             }
           }
-          let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-            '../services/clients/connector.client.js': connectorClientStub
-          })
-          const emailModel = Email('some-unique-id')
-          return expect(emailModel.setConfirmationEnabled(123, toggle))
+          const emailService = getEmailService(connectorClientStub)
+          return expect(emailService.setConfirmationEnabled(123, toggle, 'some-unique-id'))
             .to.be.rejectedWith('CONNECTOR_FAILED')
         }
         )
@@ -206,11 +189,8 @@ describe('email notification', function () {
               return sCEFinst
             }
           }
-          let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-            '../services/clients/connector.client.js': connectorClientStub
-          })
-          const emailModel = Email('some-unique-id')
-          return expect(emailModel.setConfirmationEnabled(123, true))
+          const emailService = getEmailService(connectorClientStub)
+          return expect(emailService.setConfirmationEnabled(123, true, 'some-unique-id'))
             .to.be.rejectedWith('PATCH_FAILED')
         })
       })
@@ -230,11 +210,8 @@ describe('email notification', function () {
               return sCEFinst
             }
           }
-          let Email = proxyquire(path.join(__dirname, '/../../../app/models/email.js'), {
-            '../services/clients/connector.client.js': connectorClientStub
-          })
-          const emailModel = Email('some-unique-id')
-          return expect(emailModel.setConfirmationEnabled(123, true)).to.be.fulfilled
+          const emailService = getEmailService(connectorClientStub)
+          return expect(emailService.setConfirmationEnabled(123, true, 'some-unique-id')).to.be.fulfilled
         })
       })
     })
